added contracts, changes for deploying them, enhanced role based access, added company details components

+++ backend-api/src/main/java/floread/backendapi/controller/CompanyContractCompanyController.java
@@ -22,6 +22,7 @@ import org.springframework.http.MediaType;
 import floread.backendapi.dao.AppUserDAO;
 import floread.backendapi.dao.CompanyContractCompanyDAO;
 import floread.backendapi.dao.CompanyDAO;
+import floread.backendapi.dao.RoleTypeDAO;
 import floread.backendapi.dao.UserRoleDAO;
 import floread.backendapi.entities.AppUser;
 import floread.backendapi.entities.Company;
@@ -40,6 +41,8 @@ class CompanyContractCompanyController {
     AppUserDAO appUserDAO;
     @Autowired
     UserRoleDAO userRoleDAO;
+    @Autowired
+    RoleTypeDAO roleTypeDAO;
 
     @GetMapping
     public ResponseEntity<List<CompanyContractCompany>> getAll(Principal principal) {
@@ -70,13 +73,19 @@ class CompanyContractCompanyController {
 
     @PostMapping
     public ResponseEntity<CompanyContractCompany> create(@RequestBody CompanyContractCompany item, Principal principal) {
+        
+        //this breaks if the role types are not added to db     
+        final String ROLETYPEID_ADMIN = roleTypeDAO.findByRoleCode("ADMIN").get().getRoleTypeId();
+        final String ROLETYPEID_HR = roleTypeDAO.findByRoleCode("HR").get().getRoleTypeId();
+
         Optional<AppUser> appUser = appUserDAO.findByUsername(principal.getName());
         if (appUser.isPresent()) {
             Optional<Company> company1 = companyDAO.findByCui(item.getCompanyId1());
             Optional<Company> company2 = companyDAO.findByCui(item.getCompanyId2());
             if (company1.isPresent() && company2.isPresent()){
                 Optional<UserRole> uRole = userRoleDAO.findByCompanyIdAndAppUserId(company1.get().getCompanyId(), appUser.get().getAppUserId());
-                if (uRole.isPresent() && uRole.get().getRoleType().getRoleCode() == "ADMIN" || uRole.get().getRoleType().getRoleCode() == "HR"){
+                if (!uRole.isPresent()) {return new ResponseEntity<>(HttpStatus.UNAUTHORIZED); }
+                if (uRole.get().getRoleTypeId().equals(ROLETYPEID_ADMIN) || uRole.get().getRoleTypeId().equals(ROLETYPEID_HR)){
                     try {
                         item.setCompanyId1(company1.get().getCompanyId());
                         item.setCompanyId2(company2.get().getCompanyId());
@@ -86,32 +95,67 @@ class CompanyContractCompanyController {
                         return new ResponseEntity<>(null, HttpStatus.EXPECTATION_FAILED);
                     }
                 }
+                else {
+                    return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);}
             }
             else return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
         }
         return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
     }
 
+
+    //only someone that represents company2 can accept 
     @PutMapping(value = "/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<CompanyContractCompany> update(@PathVariable("id") String id, @RequestBody CompanyContractCompany item) {
+    public ResponseEntity<CompanyContractCompany> update(@PathVariable("id") String id, @RequestBody CompanyContractCompany item, Principal principal) {
+       //this breaks if the role types are not added to db     
+        final String ROLETYPEID_ADMIN = roleTypeDAO.findByRoleCode("ADMIN").get().getRoleTypeId();
+        final String ROLETYPEID_HR = roleTypeDAO.findByRoleCode("HR").get().getRoleTypeId();
+
+        Optional<AppUser> appUser = appUserDAO.findByUsername(principal.getName());
         Optional<CompanyContractCompany> existingItemOptional = repository.findById(id);
         if (existingItemOptional.isPresent()) {
-            CompanyContractCompany existingItem = existingItemOptional.get();
-            System.out.println("TODO for developer - update logic is unique to entity and must be implemented manually.");
-            //existingItem.setSomeField(item.getSomeField());
-            return new ResponseEntity<>(repository.save(existingItem), HttpStatus.OK);
+            if (appUser.isPresent()) {
+                Optional<Company> company2 = companyDAO.findByCui(item.getCompanyId2());
+                if (company2.isPresent()){
+                    Optional<UserRole> uRole = userRoleDAO.findByCompanyIdAndAppUserId(company2.get().getCompanyId(), appUser.get().getAppUserId());
+                    if (!uRole.isPresent()) {return new ResponseEntity<>(HttpStatus.UNAUTHORIZED); }
+                    if (uRole.get().getRoleTypeId().equals(ROLETYPEID_ADMIN) || uRole.get().getRoleTypeId().equals(ROLETYPEID_HR)){
+                        CompanyContractCompany existingItem = existingItemOptional.get();
+                        existingItem.setAccepted(item.getAccepted());
+                        return new ResponseEntity<>(repository.save(existingItem), HttpStatus.OK);
+                    } else return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+                } return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            } else return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
         } else {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
     }
 
     @DeleteMapping(value = "/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<HttpStatus> delete(@PathVariable("id") String id) {
-        try {
-            repository.deleteById(id);
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
-        } catch (Exception e) {
+    public ResponseEntity<HttpStatus> delete(@PathVariable("id") String id, Principal principal) {
+         //this breaks if the role types are not added to db     
+         final String ROLETYPEID_ADMIN = roleTypeDAO.findByRoleCode("ADMIN").get().getRoleTypeId();
+         final String ROLETYPEID_HR = roleTypeDAO.findByRoleCode("HR").get().getRoleTypeId();
+ 
+        Optional<AppUser> appUser = appUserDAO.findByUsername(principal.getName());
+        CompanyContractCompany item = repository.findById(id).get();
+        if (appUser.isPresent()) {
+            Optional<Company> company1 = companyDAO.findByCui(item.getCompanyId1());
+            if (company1.isPresent()){
+                Optional<UserRole> uRole = userRoleDAO.findByCompanyIdAndAppUserId(company1.get().getCompanyId(), appUser.get().getAppUserId());
+                if (!uRole.isPresent()) {return new ResponseEntity<>(HttpStatus.UNAUTHORIZED); }
+                if (uRole.get().getRoleTypeId().equals(ROLETYPEID_ADMIN) || uRole.get().getRoleTypeId().equals(ROLETYPEID_HR)){
+                    try {
+                        repository.deleteById(id);
+                        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+                    } catch (Exception e) {
+                        return new ResponseEntity<>(HttpStatus.EXPECTATION_FAILED);
+                    }
+                }
+                else return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+            }
             return new ResponseEntity<>(HttpStatus.EXPECTATION_FAILED);
         }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 }
