+++ backend-api/src/main/java/floread/backendapi/controller/PersonWalletController.java
@@ -1,5 +1,6 @@
 package floread.backendapi.controller;
 
+import java.security.Principal;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -18,7 +19,11 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.http.MediaType;
 
+import floread.backendapi.dao.AppUserDAO;
+import floread.backendapi.dao.PersonDAO;
 import floread.backendapi.dao.PersonWalletDAO;
+import floread.backendapi.entities.AppUser;
+import floread.backendapi.entities.Person;
 import floread.backendapi.entities.PersonWallet;
 
 @RestController
@@ -27,6 +32,8 @@ class PersonWalletController {
 
     @Autowired
     PersonWalletDAO repository;
+    @Autowired
+    AppUserDAO appUserDAO;
 
     @GetMapping
     public ResponseEntity<List<PersonWallet>> getAll() {
@@ -56,35 +63,58 @@ class PersonWalletController {
     }
 
     @PostMapping
-    public ResponseEntity<PersonWallet> create(@RequestBody PersonWallet item) {
-        try {
-            PersonWallet savedItem = repository.save(item);
-            return new ResponseEntity<>(savedItem, HttpStatus.CREATED);
-        } catch (Exception e) {
-            return new ResponseEntity<>(null, HttpStatus.EXPECTATION_FAILED);
+    public ResponseEntity<PersonWallet> create(@RequestBody PersonWallet item, Principal principal) {
+        if (isUserAllowed(principal, item)){
+            try {
+                PersonWallet savedItem = repository.save(item);
+                return new ResponseEntity<>(savedItem, HttpStatus.CREATED);
+            } catch (Exception e) {
+                return new ResponseEntity<>(null, HttpStatus.EXPECTATION_FAILED);
+            }
         }
+        else return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 
     @PutMapping(value = "/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<PersonWallet> update(@PathVariable("id") String id, @RequestBody PersonWallet item) {
+    public ResponseEntity<PersonWallet> update(@PathVariable("id") String id, @RequestBody PersonWallet item, Principal principal) {
         Optional<PersonWallet> existingItemOptional = repository.findById(id);
-        if (existingItemOptional.isPresent()) {
-            PersonWallet existingItem = existingItemOptional.get();
-            System.out.println("TODO for developer - update logic is unique to entity and must be implemented manually.");
-            //existingItem.setSomeField(item.getSomeField());
-            return new ResponseEntity<>(repository.save(existingItem), HttpStatus.OK);
+        if (existingItemOptional.isPresent()){
+            if (isUserAllowed(principal, item)){
+                PersonWallet existingItem = existingItemOptional.get();
+                existingItem.setWalletCode(item.getWalletCode());
+                existingItem.setWalletDescription(item.getWalletDescription());
+                return new ResponseEntity<>(repository.save(existingItem), HttpStatus.OK);
+            }
+            else return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+            
         } else {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
     }
 
     @DeleteMapping(value = "/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<HttpStatus> delete(@PathVariable("id") String id) {
-        try {
-            repository.deleteById(id);
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
-        } catch (Exception e) {
-            return new ResponseEntity<>(HttpStatus.EXPECTATION_FAILED);
+    public ResponseEntity<HttpStatus> delete(@PathVariable("id") String id, Principal principal) {
+        Optional<PersonWallet> item = repository.findById(id);
+        if (item.isPresent()){
+            if (isUserAllowed(principal, item.get())){
+                try {
+                    repository.deleteById(id);
+                    return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+                } catch (Exception e) {
+                    return new ResponseEntity<>(HttpStatus.EXPECTATION_FAILED);
+                }
+            }
+            else return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
         }
+        else return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+       
+    }
+    
+    private Boolean isUserAllowed(Principal principal, PersonWallet item){
+        Optional<AppUser> appUser = appUserDAO.findByUsername(principal.getName());
+        if (appUser.isPresent() && appUser.get().getAppUserId() == item.getAppUserId()) {    
+            return true;
+        }
+        return false;
     }
 }
